+++ website/src/components/flex/FlexHero.js
@@ -20,12 +20,16 @@ const FlexHero = (props) => {
       {props.heroImage?.sourceUrl && (
         <div className={styles['hero-image']}>
           <Image
+            alt=""
             src={props.heroImage.sourceUrl}
             width={props.heroImage.mediaDetails.width}
             height={props.heroImage.mediaDetails.height}
             priority={props.index === 0 || props.index === 1}
-            layout="responsive"
-          />
+            sizes="100vw"
+            style={{
+              width: '100%',
+              height: 'auto'
+            }} />
         </div>
       )}
     </section>