+++ website/src/lib/wordpress/index.js
@@ -1,10 +1,21 @@
 import fetch from 'isomorphic-unfetch';
-import { WORDPRESS_API_URL } from 'lib/constants';
+import { getSettings } from 'lib/getSettings';
+import { trimLeadingSlash } from 'lib/utils';
 import { queryContent } from './graphql/queryContent';
 import { queryGlobals } from './graphql/queryGlobals';
 import { queryPosts } from './graphql/queryPosts';
 
-async function fetchAPI(query, { variables } = {}, token) {
+async function fetchAPI({
+  project = '',
+  query = {},
+  variables = {},
+  token,
+}) {
+  const SETTINGS = getSettings({ project });
+  let wordpress_api_url =
+    process.env.NEXT_PUBLIC_WORDPRESS_API_URL ||
+    process.env.WORDPRESS_API_URL ||
+    SETTINGS.CONFIG.wordpress_api_url;
   const headers = { 'Content-Type': 'application/json' };
 
   if (variables?.preview && token) {
@@ -18,7 +29,7 @@ async function fetchAPI(query, { variables } = {}, token) {
   // console.log('-------');
   // console.log('variables', variables);
   // console.log('query', typeof query, query);
-  const res = await fetch(WORDPRESS_API_URL, {
+  const res = await fetch(wordpress_api_url, {
     method: 'POST',
     headers,
     body: JSON.stringify({
@@ -37,46 +48,22 @@ async function fetchAPI(query, { variables } = {}, token) {
 }
 
 /**
- * Get post types and paths for preview mode redirects.
- */
-export async function getContentTypes(token) {
-  const query = /* GraphQL */ `
-    query ContentTypes {
-      contentTypes {
-        nodes {
-          name
-          restBase
-        }
-      }
-    }
-  `;
-
-  const data = await fetchAPI(
-    query,
-    {
-      variables: { preview: true },
-    },
-    token,
-  );
-
-  return data?.contentTypes?.nodes;
-}
-
-/**
  * To assist with Preview Mode, this will grab status for content by DB id
  * (needed for revisions, unpublished content)
  */
-export async function getPreviewContent(
+export async function getPreviewContent({
+  project,
   id,
   idType = 'DATABASE_ID',
   token,
-) {
+}) {
   const query = /* GraphQL */ `
     query PreviewContent($id: ID!, $idType: ContentNodeIdTypeEnum!) {
       contentNode(id: $id, idType: $idType) {
         uri
         status
         databaseId
+        previewRevisionDatabaseId
         contentType {
           node {
             name
@@ -87,13 +74,12 @@ export async function getPreviewContent(
     }
   `;
 
-  const data = await fetchAPI(
+  const data = await fetchAPI({
+    project,
     query,
-    {
-      variables: { id, idType, preview: true },
-    },
+    variables: { id, idType, preview: true },
     token,
-  );
+  });
 
   return data.contentNode;
 }
@@ -105,7 +91,10 @@ export async function getPreviewContent(
  * If a contentType is passed, the allQuery graphql is modified to query for
  * only that post type instead of getting posts from any CPT
  */
-export async function getAllContentWithSlug(contentType) {
+export async function getAllContentWithSlug({
+  project,
+  contentType,
+}) {
   const query = /* GraphQL */ `
     ${
       contentType
@@ -125,7 +114,9 @@ export async function getAllContentWithSlug(contentType) {
     }
   `;
 
-  const data = await fetchAPI(query, {
+  const data = await fetchAPI({
+    project,
+    query,
     variables: {
       contentType,
     },
@@ -134,55 +125,68 @@ export async function getAllContentWithSlug(contentType) {
 }
 
 /**
+ * Determine if the given pathname is a contentnode before attempting to query for it
+ * @param {*} slug pathname from URL
+ * @returns
+ */
+export async function getNodeType({ project, slug }) {
+  const query = /* GraphQL */ `
+    query getNodeType($slug: String!) {
+      nodeByUri(uri: $slug) {
+        isContentNode
+        isTermNode
+      }
+    }
+  `;
+
+  const data = await fetchAPI({
+    project,
+    query,
+    variables: { slug },
+  });
+
+  return data;
+}
+
+/**
  * Get fields for single page regardless of post type.
  */
-export async function getContent(slug, preview, previewData) {
+export async function getContent({
+  project,
+  slug,
+  preview,
+  previewData,
+  options = {},
+}) {
   let draft = false;
   if (preview) {
-    // Get the content types to help build preview URLs
-    const contentTypesArray = await getContentTypes(
-      previewData?.token,
-    );
-    const contentTypes = {};
+    // This is based on Next.js wordpress example: https://github.com/vercel/next.js/blob/canary/examples/cms-wordpress/lib/api.ts#L105-L112
+    // If the preview is for a draft or a revision, we need to query by ID.
+    // We check to see if the preview data is for the same post as the slug
+    // and if it is, we set `slug` to the ID of the post and the `uriType` to `DATABASE_ID`.
+    const postPreview = previewData?.post;
+    const trimSlug = trimLeadingSlash(slug);
+    const isId = Number.isInteger(Number(trimSlug));
+    const isSamePost = isId && Number(trimSlug) === postPreview?.id;
+    const isDraft = isSamePost && postPreview?.status === 'draft';
+    const isRevision =
+      isSamePost && postPreview?.status === 'publish';
 
-    for (const contentType of contentTypesArray) {
-      contentTypes[contentType.restBase] = contentType.name;
-    }
-
-    // Because static pages can't support query strings and those
-    // make preview mode much easier across published statuses and post types
-    // this coerces any numeric slug in preview mode to an ID and assumes post type
-    const segments = slug.split('/');
-    const lastSegment = segments[segments.length - 1];
-    const secondLastSegment =
-      segments.length > 2 ? segments[segments.length - 2] : null;
-    // Get post type URL segment
-    const postType = contentTypes[secondLastSegment];
-
-    // wordpress requires a different slug structure for various post types
-    if (slug !== '/' && !isNaN(Number(lastSegment))) {
-      draft = true;
-      if (postType === 'post') {
-        slug = `/?p=${lastSegment}`;
-      } else if (secondLastSegment) {
-        slug = `/?id=${lastSegment}&post_type=${postType}`;
-      } else {
-        slug = `/?page_id=${lastSegment}`;
-      }
-    } else if (slug !== '/') {
-      slug += '?preview=true';
+    if ((isDraft || isRevision) && postPreview?.id) {
+      draft = isDraft;
+      slug = postPreview.id;
+      options.uriType = 'DATABASE_ID';
     }
   }
 
-  let query = queryContent(draft);
+  let query = queryContent(draft, options);
 
-  const data = await fetchAPI(
+  const data = await fetchAPI({
+    project,
     query,
-    {
-      variables: { slug, preview: !!preview },
-    },
-    previewData?.token,
-  );
+    variables: { slug, preview: !!preview },
+    token: previewData?.token,
+  });
 
   return data;
 }
@@ -192,6 +196,7 @@ export async function getContent(slug, preview, previewData) {
  */
 
 export async function getPosts({
+  project,
   ids,
   first = 12,
   after = null,
@@ -205,6 +210,8 @@ export async function getPosts({
   let query = queryPosts(taxonomyType, taxonomyTerms);
 
   const data = await fetchAPI(query, {
+    project,
+    query,
     variables: {
       contentTypes,
       first,
@@ -223,7 +230,7 @@ export async function getPosts({
 
 /** All Taxonomy Terms */
 
-export async function getCategories() {
+export async function getCategories({ project }) {
   let query = /* GraphQL */ `
     query AllCategories {
       categories {
@@ -237,15 +244,15 @@ export async function getCategories() {
     }
   `;
 
-  const data = await fetchAPI(query);
+  const data = await fetchAPI({ project, query });
   return data.categories;
 }
 
 /**
  * Global Props
  * */
-export async function getGlobalProps() {
+export async function getGlobalProps({ project }) {
   let query = queryGlobals;
-  const data = await fetchAPI(query);
+  const data = await fetchAPI({ project, query });
   return data;
 }