+++ wordpress/_init/init.sh
@@ -21,6 +21,8 @@ rm wp-content/plugins/wp-stateless/vendor/symfony/console/Resources/bin/hiddenin
 # init files if they don't exist
 mkdir -p wp-content/mu-plugins
 cp -n _init/local-plugins.php wp-content/mu-plugins/local-plugins.php 2>/dev/null || :
+cp -n _init/docker-compose.local.yml docker-compose.local.yml 2>/dev/null || :
+cp -n _init/local.sql _data/local.sql 2>/dev/null || :
 # mv _init/README.md README.md 2>/dev/null || :
 
 # WP permissions