+++ wordpress/wp-content/themes/headless/functions.php
@@ -7,10 +7,9 @@
 // Load WP Config files
 //
 
-$site_name = '';
-$headless_domain = ''; // leave blank;
+// Customize these variables per site
 $staging_wp_host = 'bubsnexts.wpengine.com';
-$local_domain = 'http://localhost:3000';
+$dashboard_cleanup = false; // Optionally will hide all but our custom widget
 $docs_link = ''; // set to a path if you have a site/document for editor instructions
 
 // stellate config
@@ -23,36 +22,32 @@ $stellate_development_service_name = "";
 $stellate_development_token = "";
 
 // Determine the hosting environment we're in
-if ( defined('WP_ENV') && WP_ENV == "development" ) {
-  define('WP_HOST', 'localhost');
-  $headless_domain = $local_domain;
+if (defined('WP_ENV') && WP_ENV == 'development') {
+    define('WP_HOST', 'localhost');
+    $headless_domain = $local_domain || 'http://localhost:3000';
 } else {
-  $headless_domain = rtrim(get_theme_mod('headless_preview_url'), '/');
+    $headless_domain = rtrim(get_theme_mod('headless_preview_url'), '/');
 
-  if ( strpos($_SERVER['HTTP_HOST'], $staging_wp_host) !== false ) {
-    define('WP_HOST', 'staging');
-  } else {
-    define('WP_HOST', 'production');
-  }
+    if (strpos($_SERVER['HTTP_HOST'], $staging_wp_host) !== false) {
+        define('WP_HOST', 'staging');
+    } else {
+        define('WP_HOST', 'production');
+    }
 }
 
 // Theme Options
-function bubs_theme_options($wp_customize)
-{
-  include_once 'setup/theme-options/headless.php';
-  $wp_customize->remove_section('custom_css');
+function bubs_theme_options($wp_customize) {
+    include_once 'setup/theme-options/headless.php';
+    $wp_customize->remove_section('custom_css');
 }
 
 add_action('customize_register', 'bubs_theme_options');
 
-
 // Post Types
 //include_once 'setup/post-types/press.php';
 
-
 // Taxonomies
 //include_once 'setup/taxonomies/issue-areas.php';
-
 
 // WP Helper Functions
 
@@ -60,7 +55,8 @@ include_once 'setup/helpers/acf-options.php';
 include_once 'setup/helpers/admin.php';
 include_once 'setup/helpers/admin-env.php';
 include_once 'setup/helpers/auth.php';
-include_once 'setup/helpers/dashboard-preview.php';
+include_once 'setup/helpers/dashboard-customize.php';
+include_once 'setup/helpers/gutenberg-disable.php';
 include_once 'setup/helpers/headless-redirect.php';
 include_once 'setup/helpers/images.php';
 include_once 'setup/helpers/menus.php';
@@ -68,7 +64,7 @@ include_once 'setup/helpers/menus.php';
 // Only load permalinks rewriting file if not on a sitemap
 $uri = $_SERVER['REQUEST_URI'];
 if (strpos($uri, 'sitemap.xml') == false) {
-  include_once 'setup/helpers/permalinks.php';
+    include_once 'setup/helpers/permalinks.php';
 }
 
 include_once 'setup/helpers/role-super-editor.php';