+++ wordpress/wp-content/themes/headless/setup/helpers/menus.php
@@ -4,15 +4,15 @@
 // Add Menus to Wordpress
 //
 
-function bubs_register_nav_menu(){
-  register_nav_menus(array(
-    'header' => 'Header Navigation',
-    'footer' => 'Footer Navigation',
-    'footer_secondary' => 'Footer Secondary Navigation',
-    'footer_social' => 'Social',
-  ));
+function bubs_register_nav_menu() {
+    register_nav_menus([
+        'header' => 'Header Navigation',
+        'footer' => 'Footer Navigation',
+        'footer_secondary' => 'Footer Secondary Navigation',
+        'footer_social' => 'Social',
+    ]);
 }
 
-add_action( 'after_setup_theme', 'bubs_register_nav_menu', 0 );
+add_action('after_setup_theme', 'bubs_register_nav_menu', 0);
 
 ?>